# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume-JD Matcher API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
  email?: string | null;
}

export interface GetCurrentAuthUserResponse {
  authenticated: boolean;
  user?: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  state: string;
  codeVerifier: string;
  nonce: string;
  redirectUri: string;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  sessionToken: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteAnalysisResponse {
  success: boolean;
}

export interface CreateAnalysisBody {
  /** Full job description text */
  jobDescription: string;
  /** Job title for display */
  jobTitle?: string;
  /** Company name */
  companyName?: string;
}

export type AnalysisSummaryStatus =
  (typeof AnalysisSummaryStatus)[keyof typeof AnalysisSummaryStatus];

export const AnalysisSummaryStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface AnalysisSummary {
  id: number;
  userId: string;
  jobTitle?: string | null;
  companyName?: string | null;
  status: AnalysisSummaryStatus;
  matchScore?: number | null;
  decisionHint?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ListAnalysesResponse {
  analyses: AnalysisSummary[];
}

export type AnalysisStatusResponseStatus =
  (typeof AnalysisStatusResponseStatus)[keyof typeof AnalysisStatusResponseStatus];

export const AnalysisStatusResponseStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface AnalysisStatusResponse {
  id: number;
  status: AnalysisStatusResponseStatus;
  currentStep?: string | null;
  stepIndex?: number | null;
  totalSteps?: number | null;
  error?: string | null;
}

export interface MatchStrength {
  skill: string;
  description: string;
}

export type MatchGapSeverity =
  (typeof MatchGapSeverity)[keyof typeof MatchGapSeverity];

export const MatchGapSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface MatchGap {
  skill: string;
  description: string;
  severity: MatchGapSeverity;
}

export type RecommendationItemType =
  (typeof RecommendationItemType)[keyof typeof RecommendationItemType];

export const RecommendationItemType = {
  resume_rewrite: "resume_rewrite",
  keyword: "keyword",
  interview_prep: "interview_prep",
  pitch: "pitch",
} as const;

export interface RecommendationItem {
  type: RecommendationItemType;
  content: string;
}

export type AnalysisDetailStatus =
  (typeof AnalysisDetailStatus)[keyof typeof AnalysisDetailStatus];

export const AnalysisDetailStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface AnalysisDetail {
  id: number;
  userId: string;
  jobTitle?: string | null;
  companyName?: string | null;
  jobDescription: string;
  status: AnalysisDetailStatus;
  resumeSummary?: string | null;
  jdSummary?: string | null;
  matchScore?: number | null;
  decisionHint?: string | null;
  strengths?: MatchStrength[] | null;
  gaps?: MatchGap[] | null;
  missingKeywords?: string[] | null;
  interviewFocusAreas?: string[] | null;
  recommendations?: RecommendationItem[] | null;
  currentStep?: string | null;
  stepIndex?: number | null;
  totalSteps?: number | null;
  error?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type LoginParams = {
  returnTo?: string;
};